split battle engine to separate file (what am i doing tbh)
